# mgmt/api/openapi.yaml
openapi: 3.1.0
info: { title: Nexora management API, version: 1.0.0 }
servers: [{ url: /api/v1 }]
security: [{ session: [] }, { bearer: [] }]
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: nexora_session }
    bearer: { type: http, scheme: bearer }
  parameters:
    Id:
      {
        name: id,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
  responses:
    Error:
      {
        description: error,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          description: 'Per-line problems of submitted zone data; for license_acknowledgement_required one entry per source (line 0, message "<source key>: <notice>").'
          items:
            type: object
            required: [line, message]
            properties:
              line: { type: integer }
              message: { type: string }
...
    FilterList:
      type: object
      required:
        [
          id,
          name,
          kind,
          url,
          refresh_interval_seconds,
          enabled,
          entry_count,
          invalid_line_count,
          last_error,
          stale,
          revision,
          engine_group_id,
          category_key,
          managed_by_catalog,
        ]
      properties:
        engine_group_id:
          {
            type: [string, "null"],
            format: uuid,
            description: "engine group; null applies to every group",
          }
        category_key:
          {
            type: [string, "null"],
            description: "filter category of a catalog-managed list; null for custom lists",
          }
        managed_by_catalog:
          {
            type: boolean,
            description: "mirrors a filter category catalog source; read-only (toggle it through /filter-categories)",
          }
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [block, allow] }
        url: { type: string }
        refresh_interval_seconds: { type: integer, minimum: 300 }
        enabled: { type: boolean }
        current_blob_sha256: { type: [string, "null"] }
        entry_count: { type: integer }
        invalid_line_count: { type: integer }
        last_success_at: { type: [string, "null"], format: date-time }
        last_attempt_at: { type: [string, "null"], format: date-time }
        last_error: { type: string }
        stale:
          {
            type: boolean,
            description: last attempt failed or last success older than 2 x refresh interval,
          }
        revision: { type: integer, format: int64 }
...
    ZoneEndpoint:
      type: object
      required: [address]
      properties:
        address: { type: string, description: "ip:port" }
        tsig_key_id: { type: [string, "null"], format: uuid }
    ZoneTransfer:
      type: object
      required: [allow_cidrs]
      properties:
        allow_cidrs:
          {
            type: array,
            items: { type: string },
            description: "Clients allowed to transfer the zone; empty refuses transfers.",
          }
        tsig_key_id: { type: [string, "null"], format: uuid }
    ZoneUpdatePolicy:
      type: object
      required: [tsig_key_ids]
      properties:
        tsig_key_ids:
          {
            type: array,
            items: { type: string, format: uuid },
            description: "TSIG keys allowed to send dynamic updates; empty refuses updates.",
          }
        allow_cidrs:
          {
            type: array,
            items: { type: string },
            description: "Sources allowed to send updates; empty allows any source. TSIG is always required.",
          }
    ZoneSOA:
      type: object
      required: [mname, rname, refresh, retry, expire, minimum, ttl]
      properties:
        mname: { type: string }
        rname: { type: string }
        refresh: { type: integer, format: int64 }
        retry: { type: integer, format: int64 }
        expire: { type: integer, format: int64 }
        minimum: { type: integer, format: int64 }
        ttl: { type: integer, format: int64 }
    ZoneSOAInput:
      type: object
      required: [mname, rname]
      description: Omitted or zero timers take the defaults (refresh 10800, retry 3600, expire 1209600, minimum 3600, ttl 3600) on create and keep the current values on update.
      properties:
        mname: { type: string }
        rname: { type: string }
        refresh:
          { type: integer, format: int64, minimum: 0, maximum: 2147483647 }
        retry: { type: integer, format: int64, minimum: 0, maximum: 2147483647 }
        expire:
          { type: integer, format: int64, minimum: 0, maximum: 2147483647 }
        minimum:
          { type: integer, format: int64, minimum: 0, maximum: 2147483647 }
        ttl: { type: integer, format: int64, minimum: 0, maximum: 2147483647 }
    ZoneSecondaryStatus:
      type: object
      required:
        [
          last_refresh_at,
          last_success_at,
          next_refresh_at,
          expires_at,
          expired,
          last_error,
          last_trigger,
        ]
      properties:
        last_refresh_at: { type: [string, "null"], format: date-time }
        last_success_at: { type: [string, "null"], format: date-time }
        next_refresh_at: { type: [string, "null"], format: date-time }
        expires_at: { type: [string, "null"], format: date-time }
        expired: { type: boolean }
        last_error: { type: string }
        last_trigger: { type: string }
    Zone:
      type: object
      required:
        [
          id,
          name,
          kind,
          revision,
          serial,
          default_ttl,
          soa,
          transfer,
          notify,
          update,
          primaries,
          dnssec_enabled,
          created_at,
          updated_at,
          engine_group_id,
          allow_query_cidrs,
        ]
      properties:
        allow_query_cidrs:
          {
            type: array,
            items: { type: string },
            description: "Clients allowed to query the zone; empty uses the access control authoritative_allow_cidrs.",
          }
        engine_group_id:
          {
            type: [string, "null"],
            format: uuid,
            description: "engine group; null applies to every group",
          }
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [primary, secondary] }
        revision: { type: integer, format: int64 }
        serial: { type: integer, format: int64 }
        default_ttl: { type: integer, format: int64 }
        soa: { $ref: "#/components/schemas/ZoneSOA" }
        transfer: { $ref: "#/components/schemas/ZoneTransfer" }
        notify:
          { type: array, items: { $ref: "#/components/schemas/ZoneEndpoint" } }
        update: { $ref: "#/components/schemas/ZoneUpdatePolicy" }
        primaries:
          { type: array, items: { $ref: "#/components/schemas/ZoneEndpoint" } }
        secondary_status:
          {
            $ref: "#/components/schemas/ZoneSecondaryStatus",
            description: "Present for secondary zones only.",
          }
        dnssec_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ZoneCreate:
      type: object
      required: [name, kind]
      properties:
        engine_group_id:
          {
            type: [string, "null"],
            format: uuid,
            description: "engine group; null applies to every group",
          }
        name:
          { type: string, maxLength: 255, description: "Absolute zone name." }
        kind: { type: string, enum: [primary, secondary] }
        default_ttl:
          {
            type: integer,
            format: int64,
            minimum: 0,
            maximum: 2147483647,
            description: "Default 3600 when omitted.",
          }
        soa:
          allOf: [{ $ref: "#/components/schemas/ZoneSOAInput" }]
          description: "Required for primary zones; secondary zones take the SOA of their primary."
        nameservers:
          {
            type: array,
            items: { type: string },
            description: "Primary zones: at least one apex NS target.",
          }
        primaries:
          {
            type: array,
            items: { $ref: "#/components/schemas/ZoneEndpoint" },
            description: "Secondary zones: at least one primary.",
          }
        transfer: { $ref: "#/components/schemas/ZoneTransfer" }
        notify:
          { type: array, items: { $ref: "#/components/schemas/ZoneEndpoint" } }
        update: { $ref: "#/components/schemas/ZoneUpdatePolicy" }
        allow_query_cidrs:
          {
            type: array,
            items: { type: string },
            description: "Clients allowed to query the zone; empty uses the access control authoritative_allow_cidrs.",
          }
...
    EngineLogs:
      type: object
      required: [engine_id, lines, last_seq, oldest_seq]
      properties:
        engine_id: { type: string, format: uuid }
        lines:
          type: array
          items:
            type: object
            required: [seq, time, level, message]
            properties:
              seq: { type: integer, format: int64 }
              time: { type: string, format: date-time }
              level: { type: string, enum: [error, warn, info, debug] }
              message: { type: string }
        last_seq: { type: integer, format: int64 }
        oldest_seq: { type: integer, format: int64 }
...
    AiAgentName:
      type: string
      enum:
        [
          querylog_anomalies,
          dashboard_insights,
          filter_recommendations,
          upstream_prediction,
          rollout_risk,
          threat_classification,
          capacity_forecast,
          rpz_suggestions,
        ]
    AiAgentState:
      type: object
      required:
        [
          name,
          enabled,
          interval_seconds,
          last_started_at,
          last_finished_at,
          last_outcome,
          last_error,
          next_run_at,
          running,
        ]
      properties:
        name: { $ref: "#/components/schemas/AiAgentName" }
        enabled: { type: boolean }
        interval_seconds: { type: integer, format: int64 }
        last_started_at: { type: [string, "null"], format: date-time }
        last_finished_at: { type: [string, "null"], format: date-time }
        last_outcome: { type: string }
        last_error: { type: string }
        next_run_at: { type: [string, "null"], format: date-time }
        running: { type: boolean }
    AiStatus:
      type: object
      description: The AI layer's state. It never contains the API key.
      required:
        [
          enabled,
          reason,
          model,
          endpoint_host,
          structured_output,
          budget,
          agents,
          features,
          mcp,
        ]
      properties:
        enabled: { type: boolean }
        reason:
          {
            type: string,
            enum:
              [
                "",
                not_configured,
                incomplete_configuration,
                endpoint_not_private,
              ],
          }
        model: { type: string }
        endpoint_host: { type: string }
        structured_output: { type: string }
        budget:
          type: object
          required: [day, limit_tokens, used_tokens, background_limit_tokens]
          properties:
            day: { type: string, format: date }
            limit_tokens: { type: integer, format: int64 }
            used_tokens: { type: integer, format: int64 }
            background_limit_tokens: { type: integer, format: int64 }
        agents:
          { type: array, items: { $ref: "#/components/schemas/AiAgentState" } }
        features:
          type: object
          required: [querylog_search, config_assistant, threat_check]
          properties:
            querylog_search: { type: boolean }
            config_assistant: { type: boolean }
            threat_check: { type: boolean }
        mcp:
          type: object
          required: [enabled, read_only]
          properties:
            enabled: { type: boolean }
            read_only: { type: boolean }
...
paths:
...
  /filter-lists/{id}/refresh:
    post:
      operationId: refreshFilterList
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        {
          "200":
            {
              description: refresh attempted; body is the list after the attempt,
              content:
                {
                  application/json:
                    { schema: { $ref: "#/components/schemas/FilterList" } },
                },
            },
          "404": { $ref: "#/components/responses/Error" },
        }
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      responses:
        "200":
          description: Hosted zones ordered by name.
          content:
            application/json:
              schema:
                { type: array, items: { $ref: "#/components/schemas/Zone" } }
    post:
      operationId: createZone
      tags: [zones]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ZoneCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /engines/{id}/logs:
    get:
      operationId: getEngineLogs
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - {
            name: after,
            in: query,
            description: "Only lines with a larger sequence number.",
            schema: { type: integer, format: int64, minimum: 0 },
          }
        - {
            name: level,
            in: query,
            description: "Minimum level.",
            schema: { type: string, enum: [error, warn, info, debug] },
          }
        - {
            name: q,
            in: query,
            description: "Case-insensitive substring of the message.",
            schema: { type: string, maxLength: 128 },
          }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 1000, default: 1000 },
          }
      responses:
        "200":
          {
            description: ok,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/EngineLogs" } },
              },
          }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "501": { $ref: "#/components/responses/Error" }
        "504": { $ref: "#/components/responses/Error" }
...
  /ai/status:
    get:
      operationId: getAiStatus
      tags: [ai]
      description: The AI layer's state; answers 200 with enabled false and a reason when AI is off.
      responses:
        "200":
          {
            description: ok,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/AiStatus" } },
              },
          }
        "503": { $ref: "#/components/responses/Error" }
...